requirements.txt
pyserial>3.0